requirements.txt
streamlit
google-generativeAI
python-dotenv